pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "omniadapters"
version = "25.0.0"
description = "A collection of code for omniadapters - unified interfaces for LLM providers."
authors = [{name="GAO Hongnan", email="hongnangao@gmail.com"}]
maintainers = [{name="GAO Hongnan", email="hongnangao@gmail.com"}]
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.12"
keywords = [
    "llm",
    "adapters",
    "ai",
    "openai",
    "anthropic",
    "google",
    "gemini",
    "structured-output",
    "pydantic",
    "instructor",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
    "Typing :: Typed",
]

dependencies = [
    "pydantic>=2.12.0,<2.13.0",
    "pydantic-settings>=2.11.0,<2.12.0",
    "jsonref>=1.1.0,<2.0.0",
    "instructor>=1.13.0",
]

[project.optional-dependencies]
token = ["tiktoken>=0.12.0"]
openai = ["openai>=2.8.0,<3.0.0"]
anthropic = ["anthropic>=0.72.0,<1.0.0"]
gemini = ["google-genai>=1.42.0,<1.51.0"]
pydantic-ai = ["pydantic-ai>=0.2.0"]
all = [
    "omniadapters[token]",
    "omniadapters[openai]",
    "omniadapters[anthropic]",
    "omniadapters[gemini]",
    "omniadapters[pydantic-ai]",
]
playground = [
    "pydanticonf>=1.0.0",
]

[dependency-groups]
lint = [
    "bandit>=1.8.6",
    "commitizen>=4.8.3",
    "pre-commit>=4.3.0",
    "ruff>=0.12.11",
]

type = [
    "mypy>=1.17.1",
    "pyright>=1.1.405",
    "ty>=0.0.1a26",
    "typeguard>=4.4.4",
    "types-pyyaml>=6.0.12.20250516",
    "types-psutil>=7.0.0.20250822",
]

test = [
    "coverage>=7.10.6",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-mock>=3.14.1",
    "pytest-xdist>=3.6.1",
    "pytest-cov>=6.0.0",
]

docs = [
    "jupyter-book>=1.0.4,<1.1.0",
    "myst-nb>=1.0.0,<2.0.0",
    "sphinx-copybutton>=0.5.0,<1.0.0",
    "sphinxcontrib-mermaid>=0.9.0,<2.0.0",
    "sphinx-proof>=0.1.0,<1.0.0",
    "sphinxcontrib-bibtex>=2.0.0,<3.0.0",
    "sphinx-togglebutton>=0.3.0,<1.0.0",
    "sphinx-design>=0.5.0,<1.0.0",
]

dev = [
    { include-group = "lint" },
    { include-group = "type" },
    { include-group = "test" },
    { include-group = "docs" },
]

[project.urls]
Homepage = "https://github.com/gao-hongnan/omniadapters"
Documentation = "https://github.com/gao-hongnan/omniadapters/tree/main/playbook"
Repository = "https://github.com/gao-hongnan/omniadapters"
Issues = "https://github.com/gao-hongnan/omniadapters/issues"
Changelog = "https://github.com/gao-hongnan/omniadapters/releases"

[tool.hatch.build.targets.wheel]
packages = ["omniadapters"]
artifacts = ["omniadapters/.tiktoken_cache/**"]
exclude = [
    "tests/",
    "docs/",
    "playbook/",
    "playground/",
    "examples/",
    "scripts/",
    "*.pyc",
    "__pycache__/",
    ".mypy_cache/",
    ".pytest_cache/",
    ".ruff_cache/",
    ".coverage",
    "htmlcov/",
]

[tool.hatch.build.targets.sdist]
exclude = [
    "tests/",
    "docs/",
    "playbook/",
    "playground/",
    "examples/",
    "scripts/",
    "*.pyc",
    "__pycache__/",
    ".mypy_cache/",
    ".pytest_cache/",
    ".ruff_cache/",
    ".git/",
    ".github/",
    ".venv/",
    "dist/",
    "build/",
    ".coverage",
    "htmlcov/",
    ".DS_Store",
]

[tool.uv]
managed = true

[tool.bandit]
exclude_dirs = [
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".tox",
    "venv_*",
    "venv",
    ".venv",
    "_build",
    "buck-out",
    "build",
    "dist",
]
